Persistence Logic JPA Hibernate
-Implemented the Persistence Logic to store, retrieve and delete the
survey data using JPA/Hibernate.
“Search Surveys” link, which allows users to search records based on any
combination of the last name, first name, city, or state. In addition,
the last column of all records rendered on the search result page have a
Delete hyperlink, which allows a user to delete the corresponding record
not only from the database, but also from the search result page.